chore(steps): move SauceLoginSteps.java and SauceSortingSteps.java to steps/sauce package
