+++ src/test/java/steps/sauce/SauceSortingSteps.java
@@ -1,4 +1,4 @@
-package steps;
+package steps.sauce;
 
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Then;
